+++ tests/test_cli.py
@@ -4,6 +4,21 @@ from click.testing import CliRunner
 from ritas.cli import main
 
 
+def test_cli_geotiff() -> None:
+    """Test that we can create a GeoTIFF file."""
+    runner = CliRunner()
+    result = runner.invoke(
+        main,
+        [
+            "--input",
+            "tests/data/square_100m.csv",
+            "--output",
+            "tests/data/square_100m.tiff",
+        ],
+    )
+    assert result.exit_code == 0
+
+
 def test_cli() -> None:
     """Test the CLI."""
     runner = CliRunner()