+++ CHANGELOG.md
@@ -5,9 +5,12 @@
 
 ### API Changes
 
+- Depend on `rasterio` for IO support.
+
 ### New Features
 
 - Added `--swath-width` or `-w` CLI switch to specify a width (#19).
+- Support GeoTIFF output (#14).
 
 ### Bug Fixes